@@ -11,13 +11,14 @@
     },
     "dependencies": {
         "antd": "^4.6.2",
+        "axios": "^0.20.0",
         "classnames": "^2.2.6",
         "date-fns": "^2.16.1",
         "gray-matter": "^4.0.2",
         "immer": "^7.0.9",
         "moment": "^2.29.0",
         "mongoose": "^5.10.5",
-        "next": "9.3.5",
+        "next": "9.5.5",
         "next-auth": "^3.1.0",
         "postcss-flexbugs-fixes": "^4.2.1",
         "postcss-preset-env": "^6.7.0",
@@ -26,12 +27,16 @@
         "react-dnd-html5-backend": "^11.1.3",
         "react-dom": "16.13.1",
         "react-query": "^2.23.1",
+        "react-query-devtools": "^2.6.0",
         "remark": "^12.0.1",
         "remark-html": "^12.0.0",
         "sass": "^1.26.10",
         "styled-components": "^5.2.0",
         "tailwindcss": "^1.7.6",
         "use-immer": "^0.4.1"
+    },
+    "peerOptionalDependencies": {
+        "mongodb": "^3.6.0"
     },
     "devDependencies": {
         "@typescript-eslint/eslint-plugin": "^4.1.1",
